backend/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.23",
    "langchain-google-genai>=2.1.3",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
